Add VS Code config to make rust-analyzer work
I realize that not everyone is using VS Code, but this configuration
will be helpful to anyone using it with this repository, while having
little to no impact on anyone else.

Please note that this configuration selects the Cargo feature that
matches my hardware. It might not be appropriate to others. I can't
really think of a better solution, except maybe not to add this
configuration at all.