Add and impl pwm traits to the alternate_functions module and use those in the pwm module
This requires some changes in the interface of the pwm module: Before,
`Pins` was impl'ed on macro-generated tuples covering different concrete pin
combinations. This is not possible anymore because this would mean to
impl `Pins` on different generic tuples of the same size which leads to
conflicting implementations. The attempts to do so are currently kept as
comments but should be removed in the future. The new approach requires
preparing a `Pins` struct which lets the pwm initialization function
determine at runtime which channels should be enabled.
`Pins.has_channelX` can only be set to true with a correctly configured
gpio pin (impl PwmChX + AlternatePP) so the appropriate pin
configuration has to be done before.