Use the MSI as default/fallback clock source for sysclk
On the one hand, this allows the user to also configure slower sysclk
speeds, on the other hand, this also matches ST's default.

It is still not possible to bypass the PLL when HSI or HSE should be
used as clock source for sysclk but the same frequency configurations
can be achieved by an appropriate PLL configuration which seems like a
better trade-off than not offering slower clock speeds like before.